santoni: biometrics: Add sanity check for enumerate() return value
On successful enumeration, the max fingerprint pointer would have been
modified to match the total number of fingerprints. The total number
would also be returned directly. This is in accordance with the 2.0
enumerate() behaviour, as seen in the old libhardware header:

    /*
     * Enumerate all the fingerprint templates found in the directory set by
     * set_active_group()
     ...
     * - The size of the array provided, in fingerprint_finger_id_t elements.
     * Max_size is a bi-directional parameter and returns the actual number
     * of elements copied to the caller supplied array.
     * In the absence of errors the function returns the total number of templates
     ...
     * The caller of this function has a complete list of the templates when *max_size
     * is the same as the function return.
     *
     * Function return: Total number of fingerprint templates in the current storage directory.
     *                  or a negative number in case of error, generally from the errno.h set.
     */
    int (*enumerate)(struct fingerprint_device *dev, fingerprint_finger_id_t *results,
        uint32_t *max_size);

For the case of Goodix, enumerate() is a dummy function that does
nothing. 0 would be returned and n would be untouched. However our
previous check for (ret == 0) was always true for Goodix even with
stored fingerprints, triggering the framework for a complete cleanup
for all fingerprints and deleting saved fingerprints every reboot (or
to be exact, each time the HAL (re)starts).

[BACKPORT FROM P] This change is redundant on O, since fingerprint
cleanup was only enabled in P. But this fixes the enumeration logic
anyway.